save in float32, so time axis has enough precision (and voltage?)

+++ software/HaasoscopeQt.py
@@ -689,7 +689,7 @@ class MainWindow(TemplateBaseClass):
             if d.xscale != self.oldxscale:
                 datatosave=d.xydata # since the x scale changed (or is the first event), save all data for this event
                 self.oldxscale=d.xscale
-            h5ds = self.outf.create_dataset(str(self.nevents), data=datatosave, dtype='float16', compression="lzf") #compression="gzip", compression_opts=5)
+            h5ds = self.outf.create_dataset(str(self.nevents), data=datatosave, dtype='float32', compression="lzf") #compression="gzip", compression_opts=5)
             #about 3kB per event per board (4 channels) for 512 samples
             h5ds.attrs['time']=time_s
             h5ds.attrs['trigger_position']=str(self.vline*d.xscaling)
